Simplify clock handling, using open-loop mode and rebasing on upstream
The kliment branch gets really muddy, I found the upstream to be a
better source of info. I removed the dependancies on external xtal by
while closely following the style of the upstream. It is also
simplified by running the DFLL open-loop. There isn't much to be
gained by locking to an on-board RC oscillator. Only the course
calibration value is loaded from NVM; the datasheet thinks that's
close enough for open-loop.